AP-1 - Rename component
