+++ components/salary-calculator-results.tsx
